requirements.txt
# Requirements for the Example Color Palette generator

cairosvg   # For SVG generation
Pillow     # For PNG generation
argparse   # For command-line arguments